VORTEX-5414: Moving heatmap to standalone project.
