--- open-sphere-base/analysis/src/main/java/io/opensphere/analysis/heatmap/package-info.java
@@ -1,2 +1,0 @@
-/** Client-side heat map classes. */
-package io.opensphere.analysis.heatmap;
