+++ pom.xml
@@ -1142,6 +1142,11 @@
             </dependency>
             <dependency>
                 <groupId>io.open-sphere.plugins</groupId>
+                <artifactId>heatmap</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.open-sphere.plugins</groupId>
                 <artifactId>imagery</artifactId>
                 <version>${project.version}</version>
             </dependency>
